pyproject.toml
[project]
name = "mcp-weather"
version = "0.1.0"
description = "Weather forecast tool for MCP"
authors = []
dependencies = [
    "python-dotenv",
    "flatdict",
    "aiohttp",
    "uvicorn",
    "fastapi",   # <-- add this
    "fastmcp>=2.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mcp-weather = "mcp_weather.server:main"

[tool.hatch.metadata]
allow-direct-references = true
